Make max_decompressed_section_size a usize

+++ symbolic-debuginfo/src/object.rs
@@ -131,7 +131,7 @@ pub struct ParseObjectOptions {
     /// Maximum uncompressed size for compressed debug file sections.
     ///
     /// This is only relevant to ELF objects.
-    pub max_decompressed_section_size: Option<u64>,
+    pub max_decompressed_section_size: Option<usize>,
 }
 
 /// Tries to infer the object type from the start of the given buffer.
